requirements.txt
beautifulsoup4==4.13.4
fastapi==0.115.12
passlib==1.7.4
pydantic==2.11.5
pymongo==4.13.0
python-dotenv==1.1.0
python_jose==3.4.0
Requests==2.32.3
uvicorn==0.34.2